+++ include/geometry.h
@@ -1,0 +1,7 @@
+#ifndef GEOMETRY_H
+#define GEOMETRY_H
+
+
+int points_distance(int x1, int y1, int x2, int y2);
+
+#endif // GEOMETRY_H